requirements.txt
streamlit
requests
python-dotenv
crewai
litellm